refactor : rename and move icon config
