+++ src/components/icon/config.ts
